First working version
In this first version I basically implemented the flow to connect to AirVantage API through the Oauth process.
This first push also include a basic scenario to find a system on AirVantage using an IMEI number.
Finally I started to design a custom mock server in order to cover those scenario with tests.